Handle flow failures
This update makes sure that release creation action doesn't fail as much
as possible after deploying to the remote servers. Additional error
messages will be shown at the job summary to clarify which steps might
have failed, to make it easier to diagnose and manually fix the issue(s).

In most cases, the issue would be with artefact uploading, tagging and creating a
release in Github. However, this is way easier to fix manually than
broken remote server uploads.

Signed-off-by: Alex Lourie <alex.lourie@zepben.com>